+++ src/edit-todo.js
@@ -1,8 +1,6 @@
 
-export const editTodoTask = (e) => {
+export const editTodoTask = (index, description) => {
     const localStoragetodos = JSON.parse(localStorage.getItem('todos'));
-    const id = parseInt(e.target.dataset.editid, 10);
-    const index = localStoragetodos.findIndex((todo) => todo.index === id);
-    localStoragetodos[index].description = e.target.value;
+    localStoragetodos[index].description = description;
     localStorage.setItem('todos', JSON.stringify(localStoragetodos));
 }
